Challenge flow response creation and exception handling (#57)

+++ card-auth-server-acs/src/main/java/org/freedomfinancestack/razorpay/cas/acs/service/authvalue/AuthValueGeneratorService.java
@@ -47,10 +47,7 @@ public class AuthValueGeneratorService {
                 getAuthValueGenerator(
                         Objects.requireNonNull(
                                 Network.getNetwork(
-                                        transaction
-                                                .getTransactionCardDetail()
-                                                .getNetworkCode()
-                                                .intValue())));
+                                        transaction.getTransactionCardDetail().getNetworkCode())));
 
         if (authValueGenerator == null) {
             log.error("getAuthValue() Error occurred as auth value generator is empty or null");
@@ -63,7 +60,7 @@ public class AuthValueGeneratorService {
     /**
      * Factory Method to fetch correct Auth Value generator corresponding to network
      *
-     * @param network which is {@link org.freedomfinancestack.razorpay.cas.dao.enums.Network}
+     * @param network {@link Network}
      * @return {@link AuthValueGenerator}
      */
     private AuthValueGenerator getAuthValueGenerator(@NonNull final Network network) {
